Integrate StableHLO at openxla/stablehlo@96acdcb7
PiperOrigin-RevId: 834068744

+++ stablehlo/stablehlo/transforms/StablehloBroadcastLowering.cpp
@@ -48,8 +48,7 @@ namespace stablehlo {
 namespace {
 
 DimensionInfo getDimensionInfo(Value op, mlir::RankedTensorType tensorType,
-                               TypeExtensionsAttr encoding,
-                               int64_t dim) {
+                               TypeExtensionsAttr encoding, int64_t dim) {
   if (!encoding || !mlir::ShapedType::isDynamic(tensorType.getDimSize(dim)))
     return DimensionInfo{tensorType.getDimSize(dim)};
 
@@ -73,7 +72,7 @@ FailureOr<Dimensions> getDimensions(Value op) {
 
   Dimensions dimensions;
   dimensions.reserve(tensor_type.getRank());
-  for (size_t idx = 0; idx < tensor_type.getRank(); ++idx) {
+  for (int64_t idx = 0; idx < tensor_type.getRank(); ++idx) {
     auto dimInfo = getDimensionInfo(op, tensor_type, encoding, idx);
     dimensions.push_back(dimInfo);
   }
@@ -89,7 +88,7 @@ FailureOr<Dimensions> getNumpyBroadcastShapeWithBounds(Value op,
   Dimensions result(max_rank);
 
   // Iterate from right to left (NumPy-style broadcasting)
-  for (int i = 1; i <= max_rank; ++i) {
+  for (size_t i = 1; i <= max_rank; ++i) {
     size_t a_idx = a.size() - i;
     size_t b_idx = b.size() - i;
     size_t res_idx = max_rank - i;
@@ -169,7 +168,7 @@ FailureOr<Dimensions> getNumpyBroadcastShape(OpBuilder& builder,
   if (failed(bcastShapeOrFail)) return failure();
   Dimensions bcastShape = std::move(*bcastShapeOrFail);
 
-  for (int i = 1; i < ops.size(); ++i) {
+  for (size_t i = 1; i < ops.size(); ++i) {
     Value currOp = ops[i];
     auto dims = getDimensions(currOp);
     if (failed(dims)) return failure();
@@ -249,7 +248,7 @@ FailureOr<Value> numpyBroadcastIfNeeded(OpBuilder& builder, Value input,
   //  - If input is not bounded, but target shape is bounded, broadcast to
   //    the padded shape then call SetDimensionSize to make dynamic.
   auto bcastShape = shape;
-  for (size_t i = 0; i < inputRank; ++i) {
+  for (int64_t i = 0; i < inputRank; ++i) {
     int64_t inputDimSize = inputShape[i].size;
     int64_t resultIdx = i + rankDiff;
     int64_t resultDimSize = shape[resultIdx].size;
